updated Slotcommit and u64 usage, renamed some files
