+++ bindings/nodejs/lib/types/block/context_input/context-input.ts
@@ -2,26 +2,23 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { AccountId } from '../..';
-import { u16 } from '../../utils/type_aliases';
-import { SlotCommitmentId } from '../slot';
+import { u16 } from '../../utils/type-aliases';
 
 enum ContextInputType {
+    // ToDo: commitment context input;
+
     /**
-     * The context input type of a `CommitmentContextInput`.
-     */
-    COMMITMENT = 0,
-    /**
-     * The context input type of a `BlockIssuanceCreditContextInput`.
+     * The context input kind of a `BlockIssuanceCreditContextInput`.
      */
     BLOCK_ISSUANCE_CREDIT = 1,
     /**
-     * The context input type of a `RewardContextInput`.
+     * The context input kind of a `RewardContextInput`.
      */
     REWARD = 2,
 }
 
 abstract class ContextInput {
-    readonly type: ContextInputType;
+    type: ContextInputType;
 
     constructor(type: ContextInputType) {
         this.type = type;
@@ -29,23 +26,11 @@ abstract class ContextInput {
 }
 
 /**
- * A Commitment Context Input references a commitment to a certain slot.
- */
-class CommitmentContextInput extends ContextInput {
-    readonly commitmentId: SlotCommitmentId;
-
-    constructor(commitmentId: SlotCommitmentId) {
-        super(ContextInputType.COMMITMENT);
-        this.commitmentId = commitmentId;
-    }
-}
-
-/**
- * A Block Issuance Credit (BIC) Context Input provides the VM with context for the value of
+ * A Block Issuance Credit (BIC) Input provides the VM with context for the value of
  * the BIC vector of a specific slot.
  */
 class BlockIssuanceCreditContextInput extends ContextInput {
-    readonly accountId: AccountId;
+    accountId: AccountId;
 
     constructor(accountId: AccountId) {
         super(ContextInputType.BLOCK_ISSUANCE_CREDIT);
@@ -57,7 +42,7 @@ class BlockIssuanceCreditContextInput extends ContextInput {
  * A Reward Context Input indicates which transaction Input is claiming Mana rewards.
  */
 class RewardContextInput extends ContextInput {
-    readonly index: u16;
+    index: u16;
 
     constructor(index: u16) {
         super(ContextInputType.REWARD);
@@ -65,26 +50,9 @@ class RewardContextInput extends ContextInput {
     }
 }
 
-const ContextInputDiscriminator = {
-    property: 'type',
-    subTypes: [
-        {
-            value: CommitmentContextInput,
-            name: ContextInputType.COMMITMENT as any,
-        },
-        {
-            value: BlockIssuanceCreditContextInput,
-            name: ContextInputType.BLOCK_ISSUANCE_CREDIT as any,
-        },
-        { value: RewardContextInput, name: ContextInputType.REWARD as any },
-    ],
-};
-
 export {
     ContextInputType,
     ContextInput,
-    CommitmentContextInput,
     RewardContextInput,
     BlockIssuanceCreditContextInput,
-    ContextInputDiscriminator,
 };
