+++ bindings/nodejs/lib/types/utils/index.ts
@@ -3,4 +3,4 @@
 
 export * from './bridge';
 export * from './hex-encoding';
-export * from './type_aliases';
+export * from './type-aliases';
